Fix docstring manipulation flagged by opus 4.8 by moving the concrete implementation of _TemporalNexusOperationHandler to the decorator
